delete hotel button

+++ apps/backend/src/app/modules/hotels/hotels.controller.ts
@@ -329,8 +329,10 @@ export class HotelsController {
 
   @Delete(':id')
   async deleteHotel(@Param() params): Promise<HotelDto> {
-    await this.hotelsRepository.delete(params.id);
     const hotelDelete: Hotel = await this.hotelsRepository.findOne(params.id);
+    if (!hotelDelete)
+      return {};
+    await this.hotelsRepository.delete(params.id);
 
     return {
       id: hotelDelete.id,
